move vm back into the core project
