+++ Cargo.toml
@@ -12,7 +12,6 @@ once_cell = "1.5.2"
 steel = { path = "./steel" }
 steel_derive = { path = "./steel_derive"}
 steel_repl = { path = "./steel_repl" }
-steel_vm = { path = "./steel_vm"}
 env_logger = "0.8.2"
 log = "0.4.11"
 
@@ -20,8 +19,7 @@ log = "0.4.11"
 members = [
     "steel",
     "steel_derive",
-    "steel_repl",
-    "steel_vm"
+    "steel_repl"
 ]
 
 [profile.release]
