+++ src/main.rs
@@ -2,8 +2,8 @@ extern crate steel;
 extern crate steel_derive;
 extern crate steel_repl;
 
+use steel::steel_vm::engine::Engine;
 use steel_repl::repl::repl_base;
-use steel_vm::engine::Engine;
 
 use std::env::args;
 use std::fs;
