+++ steel_repl/Cargo.toml
@@ -11,7 +11,6 @@ rustyline = "6.2.0"
 rustyline-derive = "0.2.0"
 colored = "1.9"
 steel = { path = "../steel" }
-steel_vm = { path = "../steel_vm" }
 tokio = { version = "1.3.0", features = ["signal", "rt", "rt-multi-thread", "sync"] }
 futures = "0.3.13"
 once_cell = "1.5.2"
