--- steel_vm/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "steel_vm"
-version = "0.1.0"
-authors = ["mattwparas <matthewparas2020@u.northwestern.edu>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-steel = { path = "../steel" }
-log = "0.4.11"
-futures = "0.3.13"
-
-[dev-dependencies]
-criterion = "0.3.3"
-
-[[bench]]
-name = "my_benchmark"
-harness = false
-
-# [profile.bench]
-# debug = true
-
-# [profile.release]
-# debug = true
